mover fit_parabola.py a interpolation.py
